apps/drama/requirements.txt
attrs==16.2.0
backports.functools-lru-cache==1.3
cffi==1.8.3
cryptography==1.5.2
cssselect==0.9.2
enum34==1.1.6
future==0.15.2
futures==3.0.5
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.18.3
google-api-python-client==1.5.3
google-gax==0.14.1
googleapis-common-protos==1.3.5
grpc-google-logging-v2==0.11.1
grpc-google-pubsub-v1==0.11.1
grpcio==1.0.0
grpcio-tools==1.0.0
httplib2==0.9.2
idna==2.1
ipaddress==1.0.17
lxml==3.6.4
oauth2client==3.0.0
parsel==1.0.3
ply==3.9
protobuf==3.1.0.post1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
PyMySQL==0.7.9
pyOpenSSL==16.1.0
python-dateutil==2.5.3
pytz==2016.7
queuelib==1.4.2
rsa==3.4.2
Scrapy==1.2.0
service-identity==16.0.0
simplejson==3.8.2
six==1.10.0
SQLAlchemy==1.1.0b3
Twisted==16.4.1
uritemplate==3.0.0
w3lib==1.15.0
zope.interface==4.3.2
